Fixed giftCard after merge
